refactor Halala
